+++ lesion_metrics/metrics.py
@@ -9,16 +9,17 @@ Created on: May 14, 2021
 """
 
 __all__ = [
+    "assd",
+    "avd",
+    "corr",
     "dice",
+    "iou_per_lesion",
+    "isbi15_score",
     "jaccard",
-    "ppv",
-    "tpr",
     "lfdr",
     "ltpr",
-    "avd",
-    "assd",
-    "corr",
-    "isbi15_score",
+    "ppv",
+    "tpr",
 ]
 
 from typing import List
@@ -27,92 +28,119 @@ from scipy.stats import pearsonr
 from skimage.measure import label
 
 from lesion_metrics.types import Label, NaN, Real
+from lesion_metrics.utils import bbox, to_numpy
 
 
 def dice(pred: Label, truth: Label) -> float:
-    """ dice coefficient between predicted and true binary masks """
+    """dice coefficient between predicted and true binary masks"""
     p, t = (pred > 0.0), (truth > 0.0)
     intersection = (p & t).sum()
     cardinality = p.sum() + t.sum()
     if cardinality == 0.0:
         return NaN
-    return 2 * intersection / cardinality
+    score: float = 2 * intersection / cardinality
+    return score
 
 
 def jaccard(pred: Label, truth: Label) -> float:
-    """ jaccard index (IoU) between predicted and true binary masks """
+    """jaccard index (IoU) between predicted and true binary masks"""
     p, t = (pred > 0.0), (truth > 0.0)
     intersection = (p & t).sum()
     union = (p | t).sum()
     if union == 0.0:
         return NaN
-    return intersection / union
+    score: float = intersection / union
+    return score
 
 
 def ppv(pred: Label, truth: Label) -> float:
-    """ positive predictive value (precision) btwn predicted and true binary masks """
+    """positive predictive value (precision) btwn predicted and true binary masks"""
     p, t = (pred > 0.0), (truth > 0.0)
     intersection = (p & t).sum()
     denom = p.sum()
     if denom == 0.0:
         return NaN
-    return intersection / denom
+    score: float = intersection / denom
+    return score
 
 
 def tpr(pred: Label, truth: Label) -> float:
-    """ true positive rate (sensitivity) between predicted and true binary masks """
+    """true positive rate (sensitivity) between predicted and true binary masks"""
     p, t = (pred > 0.0), (truth > 0.0)
     intersection = (p & t).sum()
     denom = t.sum()
     if denom == 0.0:
         return NaN
-    return intersection / denom
+    score: float = intersection / denom
+    return score
 
 
-def lfdr(pred: Label, truth: Label, iou_minimum: float=0.3) -> float:
-    """ lesion false discovery rate between predicted and true binary masks """
-    p, t = (pred > 0.0), (truth > 0.0)
-    cc, n = label(p, return_num=True)
-    if n == 0:
-        return NaN
-    count = 0
-    for i in range(1, n + 1):
-        if ((cc == i) & t).sum() == 0:
-            count += 1
-    return count / n
-
-
-def ltpr(pred: Label, truth: Label) -> float:
-    """ lesion true positive rate between predicted and true binary masks """
-    p, t = (pred > 0.0), (truth > 0.0)
+def iou_per_lesion(
+    target: Label,
+    other: Label,
+) -> List[float]:
+    """iou of each lesion using target as reference"""
+    t, o = (target > 0.0), (other > 0.0)
     cc, n = label(t, return_num=True)
-    if n == 0:
-        return NaN
-    count = 0
+    ious: List[float] = []
     for i in range(1, n + 1):
-        if ((cc == i) & p).sum() > 0.0:
-            count += 1
-    return count / n
+        lesion_bbox = bbox(cc == i)
+        target_lesion = t[lesion_bbox]
+        other_lesion = o[lesion_bbox]
+        ious.append(jaccard(other_lesion, target_lesion))
+    return ious
+
+
+def lfdr(pred: Label, truth: Label, iou_threshold: float = 0.0) -> float:
+    """lesion false discovery rate between predicted and true binary masks"""
+    assert iou_threshold >= 0.0
+    p, t = (pred > 0.0), (truth > 0.0)
+    p, t = to_numpy(p), to_numpy(t)
+    ious = iou_per_lesion(p, t)
+    if not ious:
+        return NaN
+    false_positives = [iou <= iou_threshold for iou in ious]
+    fp = sum(false_positives)
+    fp_plus_tp = len(false_positives)
+    score: float = fp / fp_plus_tp
+    return score
+
+
+def ltpr(pred: Label, truth: Label, iou_threshold: float = 0.0) -> float:
+    """lesion true positive rate between predicted and true binary masks"""
+    assert iou_threshold >= 0.0
+    p, t = (pred > 0.0), (truth > 0.0)
+    p, t = to_numpy(p), to_numpy(t)
+    ious = iou_per_lesion(t, p)
+    if not ious:
+        return NaN
+    true_positives = [iou > iou_threshold for iou in ious]
+    tp = sum(true_positives)
+    tp_plus_fp = len(true_positives)
+    score: float = tp / tp_plus_fp
+    return score
 
 
 def avd(pred: Label, truth: Label) -> float:
-    """ absolute volume difference between predicted and true binary masks """
+    """absolute volume difference between predicted and true binary masks"""
     p, t = (pred > 0.0), (truth > 0.0)
     numer = abs(p.sum() - t.sum())
     denom = t.sum()
     if denom == 0.0:
         return NaN
-    return numer / denom
+    score: float = numer / denom
+    return score
 
 
 def assd(pred: Label, truth: Label) -> float:
-    """ average symmetric surface difference between predicted and true binary masks """
+    """average symmetric surface difference between predicted and true binary masks"""
     raise NotImplementedError
 
 
 def corr(pred_vols: List[Real], truth_vols: List[Real]) -> float:
-    """ pearson correlation coefficient btwn list of predicted and true binary vols """
-    return pearsonr(pred_vols, truth_vols)[0]
+    """pearson correlation coefficient btwn list of predicted and true binary vols"""
+    coef: float = pearsonr(pred_vols, truth_vols)[0]
+    return coef
 
 
 def isbi15_score(pred: Label, truth: Label, reweighted: bool = True) -> float: